+++ src/main/java/plugily/projects/buildbattle/commands/arguments/game/ArenaSelectorArgument.java
@@ -42,6 +42,8 @@ import plugily.projects.buildbattle.handlers.language.LanguageManager;
 import plugily.projects.buildbattle.utils.Utils;
 
 import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.Map;
 
 
 /**
@@ -51,6 +53,7 @@ import java.util.ArrayList;
  */
 public class ArenaSelectorArgument implements Listener {
 
+  private final Map<Integer, BaseArena> arenas = new HashMap<>();
 
   public ArenaSelectorArgument(ArgumentsRegistry registry) {
     registry.getPlugin().getServer().getPluginManager().registerEvents(this, registry.getPlugin());
@@ -64,7 +67,12 @@ public class ArenaSelectorArgument implements Listener {
           return;
         }
         Inventory inventory = Bukkit.createInventory(player, Utils.serializeInt(ArenaRegistry.getArenas().size()), registry.getPlugin().getChatManager().colorMessage("Arena-Selector.Inv-Title"));
+
+        int slot = 0;
+        arenas.clear();
+
         for (BaseArena arena : ArenaRegistry.getArenas()) {
+          arenas.put(slot, arena);
           ItemStack itemStack;
           switch (arena.getArenaState()) {
             case WAITING_FOR_PLAYERS:
@@ -78,7 +86,7 @@ public class ArenaSelectorArgument implements Listener {
               break;
           }
           ItemMeta itemMeta = itemStack.getItemMeta();
-          itemMeta.setDisplayName(arena.getID());
+          itemMeta.setDisplayName(formatItem(LanguageManager.getLanguageMessage("Arena-Selector.Item.Name"), arena, registry.getPlugin()));
 
           ArrayList<String> lore = new ArrayList<>();
           for (String string : LanguageManager.getLanguageList("Arena-Selector.Item.Lore")) {
@@ -88,6 +96,7 @@ public class ArenaSelectorArgument implements Listener {
           itemMeta.setLore(lore);
           itemStack.setItemMeta(itemMeta);
           inventory.addItem(itemStack);
+          slot++;
         }
         player.openInventory(inventory);
       }
@@ -123,7 +132,8 @@ public class ArenaSelectorArgument implements Listener {
     Player player = (Player) e.getWhoClicked();
     player.closeInventory();
 
-    BaseArena arena = ArenaRegistry.getArena(e.getCurrentItem().getItemMeta().getDisplayName());
+
+    BaseArena arena = arenas.get(e.getRawSlot());
     if (arena != null) {
       ArenaManager.joinAttempt(player, arena);
     } else {